# openapi.yaml
openapi: 3.1.0
info:
  title: TIC API
  version: "0.1.0-mvp"
  description: >
    Contract for the TIC MVP backend. Enum values in this file must match
    shared/constants.json exactly — see docs/CONTRACT_CHANGE_PROCESS.md
    before editing either file.
servers:
  - url: /api/v1
paths:
...
  /employee/me/dashboard:
    get:
      summary: Current employee's dashboard summary
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeDashboard" }
...
  /employee/me/tasks:
    get:
      summary: Current employee's tasks
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
...
  /manager/employees/{employee_id}:
    get:
      summary: Single employee detail view for a manager, optionally filtered to one UTC calendar day
      security: [{ bearerAuth: [] }]
      parameters:
        - name: employee_id
          in: path
          required: true
          schema: { type: integer }
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
          description: Filter to a single UTC calendar day. Omit for all-time.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeDetail" }
        "404":
          description: Employee not found (or belongs to a different organization)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EmployeeDashboard:
      type: object
      description: >
        Reconciled with the actually-implemented response shape of
        GET /employee/me/dashboard and GET /manager/employees/{id} (previous
        versions of this schema described aggregate fields — active_time,
        workload_percentage, etc. — that the endpoint never actually
        returned; see docs/CONTRACT_CHANGE_PROCESS.md).
      properties:
        workPulse:
          type: array
          items:
            type: object
            properties:
              app: { type: string }
              start: { type: string }
              end: { type: string }
              duration: { type: string }
              type: { type: string }
              width: { type: string }
        applications:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              duration: { type: string }
              category: { type: string }
              lastActive: { type: string }
        hourlyActivity:
          type: array
          items:
            type: object
            properties:
              hour: { type: string }
              intensityPercent: { type: number }
        activeSeconds: { type: integer }
        idleSeconds: { type: integer }
        sessionCount: { type: integer }
...
    Task:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [TODO, IN_PROGRESS, COMPLETED, BLOCKED, OVERDUE] }
        priority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        complexity: { type: string, enum: [LOW, MEDIUM, HIGH] }
        estimated_hours: { type: number }
        actual_hours: { type: number }
        deadline: { type: string, format: date }
...
    EmployeeDetail:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        department: { type: string }
        status: { type: string }
        dashboard: { $ref: "#/components/schemas/EmployeeDashboard" }
...